Swaps out the journeyId regular expression for substitute [skip ci]
