+++ src/test/resources/logback-test.xml
@@ -8,7 +8,7 @@
         </encoder>
     </appender>
 
-    <logger name="tdl.client.queue.QueueClient" level="debug" additivity="false">
+    <logger name="tdl.client.queue.QueueBasedImplementationRunner" level="debug" additivity="false">
         <appender-ref ref="STDOUT" />
     </logger>
